Refactor Feature tests.
